+++ rust-spice-derive/src/lib.rs
@@ -4,3 +4,9 @@ Doc
 */
 
 extern crate proc_macro;
+use proc_macro::TokenStream;
+
+#[proc_macro]
+pub fn make_answer(_item: TokenStream) -> TokenStream {
+    "fn answer() -> u32 { 42 }".parse().unwrap()
+}